+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### Added
 
-- [#1312](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1312) Add support for `insert_all` and `upsert_all`
+- [#1315](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1315) Add support for `insert_all` and `upsert_all`
 
 ## v8.0.4